ejercicio comentado con la reflexion del ej 10

+++ M01E09/main.cpp
@@ -17,6 +17,8 @@ int main() {
     cout << "Suma: " << a + b << endl;
     cout << "Resta: " << a - b << endl;
     cout << "Multiplicacion: " << a * b << endl;
-    cout << "Division: " << a / b << endl;
+    cout << "Division: " << (double) a / b << endl;
     return 0;
+    // si el conjunto de datos es float estamos atr, salvo que sea 0
+    // si el conjunto es entero tenemos el problema del 0 en la division tambien
 }
